rename settings styles
